Fix cmp deprecation warnings, pass **kwargs to attrib
Since attrs 19.2.0, the 'cmp' argument is deprecated and shows the
following:

    fields.py:27: DeprecationWarning: The usage of `cmp` is deprecated
    and will be removed on or after 2021-06-01.  Please use `eq` and
    `order` instead.
        metadata=dict(key=key))

There was no processing of the argument other than passing it through,
and the old default value (True) is equivalent to the current default of
attrs (None, like cmp=True but without the warning)

This commit solves the problem by not passing cmp explicitly, but
letting the user pass cmp (and get the deprecation warning, if they
still rely on it and shouldn't) or pass the new arguments, eq and order.

So the behavior should remain the same, but deprecation warnings won't
be raised by related, but by the users who actually pass an explicit
True or False to cmp.